refactor: move amount formatting to TextUtils
